Se implementa subtarea 2.2
Se modulariza el codigo creando gol.h y gol.c
Se crea Makefile con las tareas all y clean
Se crea .gitignore para ignorar productos de la compilacion
