+++ README.md
@@ -3,13 +3,13 @@ Juego de la vida
 
 ## Instalación (para GNU/Linux)
 Abrimos una consola de comandos y ejecutamos lo siguiente:
-* Debian-based: `$ sudo apt install gcc`
-* Arch-based: `$ sudo pacman gcc`
-* Fedora-based: `$ sudo dnf install gcc`
+* Debian-based: `$ sudo apt-get update && sudo apt-get install build-essential`
+* Arch-based: `$ sudo pacman -Sy base-devel`
+* Fedora-based: `sudo yum update && sudo yum groupinstall "Development Tools" "Legacy Software Development"`
 
 ## Compilación y Ejecución
 
 1. Nos descargamos la carpeta del repositorio.
 2. Desde una terminal, entramos en el directorio del repositorio.
-3. Ejecutamos `$ gcc main.c -o main` para compilar.
-4. Y para ejecuta el programa hacemos `$ ./main`
+3. Ejecutamos `$ make` para compilar.
+4. Y para ejecuta el programa hacemos `$ ./exec`
